Move location gathering to a util
